[refactor] 코드리뷰 내용 수정 완료 #52
