+++ client/src/pages/HomePage.tsx
@@ -5,7 +5,7 @@ import { useAuth } from '@/hooks';
 const HomePage: React.FC = () => {
   const { accessToken } = useAuth();
 
-  return <>{accessToken ? <Redirect to="/workspace" /> : <Redirect to="/login" />}</>;
+  return <>{accessToken ? <Redirect to="/client/1" /> : <Redirect to="/login" />}</>;
 };
 
 export default HomePage;
